consumos funcionam!!

+++ src/main/java/com/primeira/appSpring/model/M_ViewLocacao.java
@@ -13,5 +13,5 @@ public interface M_ViewLocacao {
     Date getCheckout();
     int getDiarias();
 
-    int getConsumos();
+    BigDecimal getConsumos();
 }
